Cleanup, separating to two programs
