--- version.h
@@ -1,11 +1,0 @@
-/*
- * Perceptual image hash calculation tool based on algorithm descibed in
- * Block Mean Value Based Image Perceptual Hashing by Bian Yang, Fan Gu and Xiamu Niu
- *
- * Copyright 2014-2015 Commons Machinery http://commonsmachinery.se/
- * Distributed under an MIT license, please see LICENSE in the top dir.
- */
-
-#define PROGRAM_VERSION "1.0"
-#define COPYRIGHT_YEARS "2014-2015"
-#define OWNER_NAME "Commons Machinery http://commonsmachinery.se/"
